requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==1.10.9
mysql-connector-python==8.0.33
streamlit==1.35.0
pandas==2.0.2
requests==2.31.0
pytest==8.3.2